pyproject.toml
[tool.poetry]
name = "breachrx-cli"
version = "0.0.0"
description = ""
authors = ["BreachRx <support@breachrx.com>"]
packages = [
    { include = "breachrx-cli" }
]
[tool.poetry.dependencies]
python = ">=3.6,<4"
gql = "^3.0.0"
aiohttp = "^3.8.1"
python-dotenv = "^0.19.2"
click = "^8.0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
